+++ CHANGELOG.md
@@ -2,12 +2,17 @@
 
 All notable changes to **VTUinterface** will be documented in this file.
 
+## [0.682]
+
+### Changes
+* VTUinterface is now able to read pvtu files
+* add methods for returning neighbor points and their indices
+
 ## [0.681]
 
 ### Changes
 * new tool that enables spatial transformation of slices
 * more functionalities to read and interpolate cell data based on cell center points
-* VTUinterface is now able to read pvtu files
 
 ## [0.68]